Some Plugins aren't worth the trouble on Windows
In particular, the plugins I've been running experimentally.  Primarily
YouCompleteMe, which on Windows runs in gVim but not in vim, because
Python runs in gVim but not in vim.  And YouCompleteMe must be compiled
on Windows.  That's annoying.  I may enable UltiSnips in gVim on Windows
if it turns out to work.  Without the two of them, there's no point in
running supertab or vim-snippets.