Write a function to start a rebase
It's actually pretty complicated to move you current feature branch to
the tip of main.  Here's a function that
    * stashes everything (including untracked files)
    * switches to main
    * pulls (with rebase) to make sure main is up-to-date
    * then if you weren't on main to begin with, rebases your starting
      branch to the tip of main
    * finally, if anything was stashed in the first step, pops the stash

Of course you might have to type a password at the pull stage; and you
might face conflicts both during the rebase and for the stash pop.
That's why the function is called start_rebase_onto_origin_main.