Rewrite my replacement for tree
I kept calling (my old) tree command with a directory argument which
made it seem to lock up.  Of course those arguments were going on to
as-tree, not fd.  Rewrote the alias as a function that takes fd's
parameters (at least the pattern and target directory).  Reversed the
order because specifying the target directory to tree is much more
common than filtering with a pattern.