Use `find` versions of my Bash functions if no fd
If fd.bash ran, then I will already have definitions for these
functions.  So for each one, test if that name is already defined.  If
it's not, then go ahead and an write out the `find` version of the
function.

Note that `fx` now benefits from the color flag change.  Note that `fcd`
now uses `head` instead of `awk`.  Note that I finally got the magic
right to use `xargs` for `en` instead of relying on simple expansion.