Map Y in Vim to help in presentations
When I have a demo shell window shared, I can copy command lines from a
script on a non-shared screen.  I don't want to copy the newline at the
end, because then a paste into the shell would immediately execute the
command giving me no chance to explain it.  This mapping copies
everything on the line up to, but not including the newline.