Try Xonsh shell
Port over the most important of my quick, little, Bash shortcuts.  Don't
know if having this many rc.d files is recommended or if it's better to
build a xontrib with one or more of these.

Also added a plugin to Vim so .xsh files look/act right while editing.